Code dlean up. Warnings.
