+++ Directory.Packages.props
@@ -166,7 +166,7 @@
 
     <PackageVersion Include="Microsoft.CodeAnalysis.Common" Version="$(RoslynApiMinVersion)" />
     <PackageVersion Include="Microsoft.CodeAnalysis.CSharp" Version="$(RoslynApiMinVersion)" />
-    <PackageVersion Include="Microsoft.CodeAnalysis.Workspaces.MSBuild" Version="$(RoslynApiMaxVersion)" />
+    <PackageVersion Include="Microsoft.CodeAnalysis.Workspaces.MSBuild" Version="$(RoslynApiMinVersion)" />
     
   
     <!-- These are test-only dependencies for standalone projects. -->
