+++ Metalama.Framework/src/tests/Metalama.Framework.Tests.Benchmarks.4.12.0/Metalama.Framework.Tests.Benchmarks.4.12.0.csproj
@@ -1,0 +1,12 @@
+<Project ToolsVersion="Current">
+  <PropertyGroup>
+    <IsPackable>false</IsPackable>
+    <RoslynVersion>4.14.0</RoslynVersion>
+    <EngineProjectSuffix>.4.12.0</EngineProjectSuffix>
+  </PropertyGroup>
+  <ItemGroup>
+    <Compile Include="../Metalama.Framework.Tests.Benchmarks/**/*.cs" Exclude="../Metalama.Framework.Tests.Benchmarks/bin/**/*.cs;../Metalama.Framework.Tests.Benchmarks/obj/**/*.cs" />
+    <Compile Remove="**/*" Condition="'$(FormattingCode)'=='True'" />
+  </ItemGroup>
+  <Import Project="../Metalama.Framework.Tests.Benchmarks/Metalama.Framework.Tests.Benchmarks.csproj" />
+</Project>
